ODS plugin
Summary:
Each time when the plugin starts (which is also when you re-enter it, even if the switch is on in the UI), will query Unity for all counters.
There's also a button to manually refresh the list.
I added a button to filter 0 counters (since there are a lot of them). It will pulse while it's filtering (on by default).

Differential Revision: D21923242

fbshipit-source-id: 0dc54eb8b460fe68a22aa86f43a5115a70728092