Migrate to FBSystrace Server
Summary:
As title, this restores all of the functionality of the original plugin to but uses json-rpc to interact with fbsystrace in server mode.

Mostly at feature parity with the old plugins, need to update the following in my next diff:
- Better fbsystrace error surfacing
- Pipe all all output to 'Console'
- Support non-select type options

{F239554690}

I think that the authentication issues with Tracery might go away in prod - according to https://fb.workplace.com/groups/flippersupport/permalink/889080114905981/ this seems like a dev issue. But the "Open in Browser" functionality is working fine.

Reviewed By: nslingerland

Differential Revision: D21668522

fbshipit-source-id: fa3b887c7b5dedc08c62a9d579e1112d7f8e0323