Bump docusaurus-plugin-internaldocs-fb
Summary:
I found out that when I fixed the nginx caching issue a few weeks ago, the nginx config was deployed but not all hosts using it were restarted, so I manually restarted them and verified that all three pods are now using the "cache-control: -1" settings.

This is a bump to the plugin so we can detect any versions that are older than this, and refresh the page. I never shipped the [previous diff that tells the user to force refresh](https://www.internalfb.com/intern/diff/D21906069/) because I didn't fully understand what was going on, but now I do.

So I'll make a new one to refresh transparently without the user needing to know.

Reviewed By: nikoant

Differential Revision: D22159282

fbshipit-source-id: 0b7a9e21ea3ad04f99b84eedd585985c8aae232c